[219]: Renaming per PR suggestions
